requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.10
numpy==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.10
PuLP==2.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.10.1
python-jose==3.3.0
python-multipart==0.0.19
rsa==4.9
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.36.3
typing_extensions==4.12.2
uvicorn==0.27.1
